requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
idna==3.10
jinja2==3.1.6
markupsafe==3.0.2
mpmath==1.3.0
mypy-extensions==1.1.0
networkx==3.5
numpy==2.3.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pydantic==2.11.5
pydantic-core==2.33.2
pyyaml==6.0.2
setuptools==80.9.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
torch==2.7.1
typing-extensions==4.14.0
typing-inspection==0.4.1
uvicorn==0.34.3
